api.jooda.com-main/repo/apps/churchs/v1/serializers/directions_serializer.py
from rest_framework import serializers
from apps.churchs.models import ChurchDirections


class ChurchDirectionsSerializer(serializers.ModelSerializer):
    class Meta:
        model = ChurchDirections
        fields = (
            "parking",
            "own_car",
            "public_transport",
            "shuttle_bus",
        )
