api.jooda.com-main/repo/requirements.txt
asgiref==3.6.0
black==23.1.0
boto3==1.26.96
botocore==1.29.96
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-redis==5.2.0
django-storages==1.13.2
djangorestframework==3.14.0
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
msgpack==1.0.5
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.1.1
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.2
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
